application_builder/requirements.txt
asgiref==3.7.2
Django==4.2.1
django-cors-headers==4.3.1
django-multiselectfield==0.1.12
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
